requirements.txt
# requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
diffusers
transformers
opencv-python
sentencepiece
accelerate
imageio
Flask
flask-cors
loguru
python-dotenv
black
pre-commit
requests
python-dotenv
